src/pages/AcrticlesPage/model/services/fetchArticlesNextPage/fetchArticlesNextPage.ts
import { createAsyncThunk } from '@reduxjs/toolkit';
import { ThunkConfig } from 'app/providers/StoreProvider';
import { articlePageActions } from '../../slices/articlePageSlice';
import { fetchArticlesList } from '../fetchArticlesList/fetchArticlesList';
import {
    getArticlesPageHasMore,
    getArticlesPageNum,
    getArticlesPageIsLoading,
} from '../../selectors/articlesPageSelector';

export const fetchArticlesNextPage = createAsyncThunk<
    void,
    void,
    ThunkConfig<string>
>(
    'articlesPage/fetchArticlesNextPage',
    async (_, thunkAPI) => {
        const { getState, dispatch } = thunkAPI;
        const hasMore = getArticlesPageHasMore(getState());
        const page = getArticlesPageNum(getState());
        const isLoading = getArticlesPageIsLoading(getState());
        if (hasMore && !isLoading) {
            dispatch(fetchArticlesList(
                {
                    page: page + 1,
                },
            ));

            dispatch(articlePageActions.setPage(page + 1));
        }
    },
);
